[addons] rename code_generator.py -> code_generator__addon_devkit.py
This rename the script to another name, background is the use of several scripts in
the folder and to have more clear flow about script files.

Backround is the comming new script where named "code_generator__addon_depends_angle.py"
to auto generate CMake support in google/angle.
