+++ xbmc/addons/kodi-dev-kit/tools/code-generator/code_generator__addon_devkit.py
