+++ xbmc/addons/kodi-dev-kit/tools/code-generator/src/generateCMake__XBMC_ADDONS_KODIDEVKIT_INCLUDE_KODI_allfiles.py
@@ -7,7 +7,7 @@
 #   See LICENSES/README.md for more information.
 
 # Own includes
-from code_generator import DEVKIT_DIR, KODI_DIR
+from code_generator__addon_devkit import DEVKIT_DIR, KODI_DIR
 from .helper_Log import *
 
 # Global includes
